feat: add options to write post in anonymously or logged in
