--- src/components/auth/buttons/KakaoButton.tsx
@@ -1,7 +1,0 @@
-function KakaoButton() {
-    return (
-        <img src="/images/kakao_login.png" />
-    )
-}
-
-export default KakaoButton
